\name{BatchMixtureModel-package}
\alias{BatchMixtureModel-package}
\alias{BatchMixtureModel}
\docType{package}
\title{\packageTitle{BatchMixtureModel}}
\description{\packageDescription{BatchMixtureModel}}
\section{Package Content}{\packageIndices{BatchMixtureModel}}
\author{\packageAuthor{BatchMixtureModel}}
\section{Maintainer}{\packageMaintainer{BatchMixtureModel}}
\keyword{package}
