% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchMixtureModel.R
\name{batchMixtureModel}
\alias{batchMixtureModel}
\title{Batch mixture model}
\usage{
batchMixtureModel(
  X,
  R,
  thin,
  batch_vec,
  type,
  initial_labels = NULL,
  K_max = 50,
  alpha = 1,
  mu_proposal_window = 0.5^2,
  cov_proposal_window = 100,
  m_proposal_window = 0.3^2,
  S_proposal_window = 100,
  t_df_proposal_window = 100,
  phi_proposal_window = 1.2^2,
  rho = 41,
  theta = 40,
  verbose = FALSE,
  doCombinations = FALSE,
  printCovariance = FALSE
)
}
\arguments{
\item{X}{Data to cluster as a matrix with the items to cluster held in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if
``thin=50`` only every 50th sample is kept.}

\item{batch_vec}{Labels identifying which batch each item being clustered is
from.}

\item{type}{Character indicating density type to use. One of 'MVN' 
(multivariate normal distribution), 'MVT' (multivariate t distribution) or 
'MSN' (multivariate skew normal distribution).}

\item{initial_labels}{Labels to begin from (if ``NULL`` defaults to a 
stick-breaking prior).}

\item{K_max}{The number of components to include (the upper bound on the 
number of clusters found in each sample).}

\item{alpha}{The concentration parameter for the stick-breaking prior and the 
weights in the model.}

\item{mu_proposal_window}{The proposal window for the cluster mean proposal
kernel.}

\item{cov_proposal_window}{The proposal window for the cluster covariance 
proposal kernel.}

\item{m_proposal_window}{The proposal window for the batch mean proposal
kernel.}

\item{S_proposal_window}{The proposal window for the batch standard deviation
proposal kernel.}

\item{t_df_proposal_window}{The proposal window for the degrees of freedom 
for the multivariate t distribution (not used if type is not 'MVT').}

\item{phi_proposal_window}{The proposal window for the shape parameter for
the multivariate skew normal distribution (not used if type is not 'MSN').}

\item{verbose}{A bool indicating if the acceptance count for each parameter 
should be printed or not.}
}
\value{
Named list of the matrix of MCMC samples generated (each row
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
A Bayesian mixture model with batch effects.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples and BIC vector
samples <- mixtureModel(X, R, thin)

# Predicted clustering and PSM
pred_cl <- mcclust::maxpear(samples$samples)$cl
psm <- createSimilarityMatrix(pred_cl)
}
