% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchSemiSupervisedMixtureModel.R
\name{batchSemiSupervisedMixtureModel}
\alias{batchSemiSupervisedMixtureModel}
\title{Batch semi-supervised mixture model}
\usage{
batchSemiSupervisedMixtureModel(
  X,
  R,
  thin,
  initial_labels,
  fixed,
  batch_vec,
  type,
  K_max = length(unique(initial_labels)),
  alpha = NULL,
  mu_proposal_window = 0.5^2,
  cov_proposal_window = 100,
  m_proposal_window = 0.3^2,
  S_proposal_window = 100,
  t_df_proposal_window = 100,
  phi_proposal_window = 1.2^2
)
}
\arguments{
\item{X}{Data to cluster as a matrix with the items to cluster held in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if 
``thin=50`` only every 50th sample is kept.}

\item{initial_labels}{Initial clustering.}

\item{fixed}{Which items are fixed in their initial label.}

\item{batch_vec}{Labels identifying which batch each item being clustered is from.}

\item{type}{Character indicating density type to use. One of 'MVN' 
(multivariate normal distribution), 'MVT' (multivariate t distribution) or 
'MSN' (multivariate skew normal distribution).}

\item{K_max}{The number of components to include (the upper bound on the
number of clusters in each sample). Defaults to the number of unique labels 
in ``initial_labels``.}

\item{alpha}{The concentration parameter for the stick-breaking prior and the 
weights in the model.}

\item{mu_proposal_window}{The proposal window for the cluster mean proposal
kernel.}

\item{cov_proposal_window}{The proposal window for the cluster covariance 
proposal kernel.}

\item{m_proposal_window}{The proposal window for the batch mean proposal
kernel.}

\item{S_proposal_window}{The proposal window for the batch standard deviation
proposal kernel.}

\item{t_df_proposal_window}{The proposal window for the degrees of freedom 
for the multivariate t distribution (not used if type is not 'MVT').}

\item{phi_proposal_window}{The proposal window for the shape parameter for
the multivariate skew normal distribution (not used if type is not 'MSN').}
}
\description{
A Bayesian mixture model with batch effects.
}
\examples{

# Data in a matrix format
X <- matrix(c(rnorm(100, 0, 1), rnorm(100, 3, 1)), ncol = 2, byrow = TRUE)

# Initial labelling
labels <- c(rep(1, 10), 
  sample(c(1,2), size = 40, replace = TRUE), 
  rep(2, 10), 
  sample(c(1,2), size = 40, replace = TRUE)
)

fixed <- c(rep(1, 10), rep(0, 40), rep(1, 10), rep(0, 40))

# Batch
batch_vec <- sample(1:5, replace = TRUE, size = 100)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples and BIC vector
samples <- batchSemiSupervisedMixtureModel(X, R, thin, labels, fixed, batch_vec, "MVN")
}
