% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchSemiSupervisedMVN.R
\name{batchSemiSupervisedMVN}
\alias{batchSemiSupervisedMVN}
\title{Batch semi-supervised mixture model}
\usage{
batchSemiSupervisedMVN(
  X,
  R,
  thin,
  initial_labels,
  fixed,
  batch_vec,
  K_max = length(unique(initial_labels)),
  alpha = 1,
  mu_proposal_window = 0.5^2,
  cov_proposal_window = 100,
  m_proposal_window = 0.3^2,
  S_proposal_window = 100,
  rho = 41,
  theta = 40,
  lambda = 1,
  verbose = FALSE,
  doCombinations = FALSE,
  printCovariance = FALSE
)
}
\arguments{
\item{X}{Data to cluster as a matrix with the items to cluster held in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if ``thin=50`` only every 50th sample is kept.}

\item{initial_labels}{Initial clustering.}

\item{fixed}{Which items are fixed in their initial label.}

\item{batch_vec}{Labels identifying which batch each item being clustered is from.}

\item{K_max}{The number of components to include (the upper bound on the number of clusters found). Defaults to the number of unique labels in ``initial_labels``.}

\item{alpha}{The concentration parameter for the stick-breaking prior and the weights in the model.}

\item{mu_proposal_window}{The proposal window for the cluster mean proposal kernel.}

\item{cov_proposal_window}{The proposal window for the cluster covariance proposal kernel.}

\item{m_proposal_window}{The proposal window for the batch mean proposal kernel.}

\item{S_proposal_window}{The proposal window for the batch standard deviation proposal kernel.}

\item{verbose}{A bool indicating if the acceptance count for each parameter should be printed or not.}
}
\value{
Named list of the matrix of MCMC samples generated (each row
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
A Bayesian mixture model with batch effects.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples and BIC vector
samples <- mixtureModel(X, R, thin)

# Predicted clustering and PSM
pred_cl <- mcclust::maxpear(samples$samples)$cl
psm <- createSimilarityMatrix(pred_cl)
}
