% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAllocProb.R
\name{calcAllocProb}
\alias{calcAllocProb}
\title{Caclulate allocation probabilities}
\usage{
calcAllocProb(samples, burn)
}
\arguments{
\item{samples}{Output from the ``batchSemiSupervisedMixtureModel`` or
``batchMixtureModel``.}

\item{burn}{The number of samples to discard.}
}
\value{
An N x K matrix of class probabilities.
}
\description{
Calculate the empirical allocation probability for each class
based on the sampled allocations.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 10000
thin <- 50

# MCMC samples
samples <- batchMixtureModel(X, R, thin, type = "MVN")

burn <- 2000
eff_burn <- burn / thin
probs <- allocProb(samples, burn = eff_burn)
}
