% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectAcceptanceRates.R
\name{collectAcceptanceRates}
\alias{collectAcceptanceRates}
\title{Collect acceptance rate}
\usage{
collectAcceptanceRates(samples, type)
}
\arguments{
\item{type}{The type of mixture model used; this changes which parameters
the function expects to find.}
}
\value{
A wide data.frame of all the sampled parameters and the iteration.
}
\description{
Collects the acceptance rates for each parameter into a data.frame
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples
samples <- batchMixtureModel(X, R, thin, type = "MVN")

acceptance_df <- collectAcceptanceRates(samples, "MVN")
}
