% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateBatchDataVaryingRepresentation.R
\name{generateBatchDataVaryingRepresentation}
\alias{generateBatchDataVaryingRepresentation}
\title{Generate batch data}
\usage{
generateBatchDataVaryingRepresentation(
  N,
  P,
  cluster_means,
  std_dev,
  batch_shift,
  batch_var,
  class_weights,
  batch_weights,
  frac_known = 0.2,
  row_names = paste0("Person_", 1:N),
  col_names = paste0("Gene_", 1:P)
)
}
\arguments{
\item{N}{The number of items (rows) to generate.}

\item{P}{The number of columns in the generated dataset.}

\item{cluster_means}{A vector of the class means for a column.}

\item{std_dev}{A vector of class standard deviations for a column.}

\item{batch_shift}{A vector of batch means in a column.}

\item{batch_var}{A vector of batch standard deviations within a column.}

\item{batch_weights}{A vector of the expected proportion of N in each batch.}

\item{row_names}{The column names of the final data matrix. Defaults to ``paste0("Gene_", 1:P)``.}

\item{cluster_weights}{A K x B matrix of the expected proportion of N in each class in each batch.}
}
\value{
A list of 4 objects; the data generated from the clusters with and
without batch effects, the label indicating the generating cluster and the
batch label.
}
\description{
Generate data from clusters across batches. Assumes independence
across columns. In each column the parameters are randomly permuted for both
the clusters and batches.
}
\examples{
N <- 500
P <- 2
K <- 2
B <- 5
mean_dist <- 4
batch_dist <- 0.3
cluster_means <- 1:K * mean_dist
batch_shift <- rnorm(B, mean = batch_dist, sd = batch_dist)
std_dev <- rep(2, K)
batch_var <- rep(1.2, B)
cluster_weights <- matrix(c(
  0.8, 0.6, 0.4, 0.2, 0.2,
  0.2, 0.4, 0.6, 0.8, 0.8),
  nrow = K, ncol = B, byrow = T
)
batch_weights <- rep(1 / B, B)

my_data <- generateBatchData(
  N,
  P,
  cluster_means,
  std_dev,
  batch_shift,
  batch_var,
  cluster_weights,
  batch_weights
)
}
