% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampledBatchScale.R
\name{getSampledBatchScale}
\alias{getSampledBatchScale}
\title{Get sampled batch shift}
\usage{
getSampledBatchScale(
  sampled_batch_scale,
  B = dim(sampled_batch_scale)[2],
  P = dim(sampled_batch_scale)[1],
  R = dim(sampled_batch_scale)[3],
  thin = 1
)
}
\arguments{
\item{B}{The number of batches present. Defaults to the number of columns in
the batch mean matrix from the first sample.}

\item{P}{The dimension of the batch mean shifts. Defaults to the number of 
rows in the batch mean matrix from the first sample.}

\item{R}{The number of iterations run. Defaults to the number of slices in
the sampled batch mean array.}

\item{thin}{The thinning factor of the sampler. Defaults to 1.}

\item{sampled_batch_shift}{A 3D array of sampled batch mean shifts.}
}
\value{
A data.frame of three columns; the parameter, the sampled value and the iteration.
}
\description{
Given an array of sampled batch scales from the 
``mixtureModel`` function, acquire a tidy version ready for ``ggplot2`` use.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples
samples <- mixtureModel(X, R, thin)

batch_shift_df <- getSampledBatchShift(samples$batch_shift, R = R, thin = thin)
}
