% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixtureModel.R
\name{mixtureModel}
\alias{mixtureModel}
\title{Mixture model}
\usage{
mixtureModel(
  X,
  R,
  thin,
  batch_vec,
  initial_labels = NULL,
  K_max = 50,
  alpha = 1,
  mean_proposal_window = 0.5^2,
  cov_proposal_window = 100,
  S_proposal_window = 100,
  verbose = FALSE
)
}
\arguments{
\item{X}{Data to cluster as a matrix with the items to cluster held in rows.}

\item{R}{The number of iterations in the sampler.}

\item{thin}{The factor by which the samples generated are thinned, e.g. if ``thin=50`` only every 50th sample is kept.}

\item{batch_vec}{Labels identifying which batch each item being clustered is from.}

\item{initial_labels}{Labels to begin from (if ``NULL`` defaults to a stick-breaking prior).}

\item{K_max}{The number of components to include (the upper bound on the number of clusters found).}

\item{alpha}{The concentration parameter for the stick-breaking prior and the weights in the model.}

\item{verbose}{A bool indicating if the acceptance count for each parameter should be printed or not.}

\item{mean_proposal_windows}{The proposal window for the mean proposal kernel.}

\item{cov_proposal_windows}{The proposal window for the cluster covariance proposal kernel.}

\item{S_proposal_windows}{The proposal window for the batch standard deviation proposal kernel.}
}
\value{
Named list of the matrix of MCMC samples generated (each row
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
A Bayesian mixture model.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples and BIC vector
samples <- mixtureModel(X, R, thin)

# Predicted clustering and PSM
pred_cl <- mcclust::maxpear(samples$samples)$cl
psm <- createSimilarityMatrix(pred_cl)
}
