% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampledBatchMeans.R
\name{plotSampledBatchMeans}
\alias{plotSampledBatchMeans}
\title{Plot sampled batch means}
\usage{
plotSampledBatchMeans(samples, R = NULL, thin = 1)
}
\arguments{
\item{samples}{The output of the ``mixtureModel`` function.}

\item{R}{The number of iterations run. Defaults to the number of samples for
the cluster membership.}

\item{thin}{The thinning factor of the sampler. Defaults to 1.}
}
\value{
A ggplot object of the values in each sampled batch mean per iteration.
}
\description{
Plot the sampled values for the batch mean shifts in each 
dimension from the output of the ``mixtureModel`` function. Not recommended
for large B or P.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples
samples <- mixtureModel(X, R, thin)
plotSampledBatchMeans(samples, R, thin)
}
