% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampledClusterMeans.R
\name{plotSampledClusterMeans}
\alias{plotSampledClusterMeans}
\title{Plot sampled cluster means}
\usage{
plotSampledClusterMeans(samples, R = NULL, thin = 1, burn_in = 0)
}
\arguments{
\item{samples}{The output of the ``mixtureModel`` function.}

\item{R}{The number of iterations run. Defaults to the number of samples for
the cluster membership.}

\item{thin}{The thinning factor of the sampler. Defaults to 1.}
}
\value{
A ggplot object of the values in each sampled cluster mean per iteration.
}
\description{
Plot the sampled values for the cluster means in each
dimension from the output of the ``mixtureModel`` function. Not recommended
for large K or P.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 1000
thin <- 50

# MCMC samples
samples <- mixtureModel(X, R, thin)
plotSampledClusterMeans(samples, R, thin)
}
