% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictClass.R
\name{predictClass}
\alias{predictClass}
\title{Predict class}
\usage{
predictClass(prob)
}
\arguments{
\item{prob}{Output from the ``calcAllocProb`` function, a N x K matrix of 
allocation probabilities.}
}
\value{
An N vecotr of class allocations.
}
\description{
Predicts a final class for each item given a matrix of 
allocation probabilities.
}
\examples{
# Convert data to matrix format
X <- as.matrix(my_data)

# Sampling parameters
R <- 10000
thin <- 50

# MCMC samples
samples <- batchMixtureModel(X, R, thin, type = "MVN")

burn <- 2000
eff_burn <- burn / thin
probs <- allocProb(samples, burn = eff_burn)
preds <- predictClass(probs)
}
