% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sampleMVN}
\alias{sampleMVN}
\title{Sample batch mixture model}
\usage{
sampleMVN(
  X,
  K,
  B,
  labels,
  batch_vec,
  mu_proposal_window,
  cov_proposal_window,
  m_proposal_window,
  S_proposal_window,
  rho,
  theta,
  lambda,
  R,
  thin,
  concentration,
  verbose = TRUE,
  doCombinations = FALSE,
  printCovariance = FALSE
)
}
\arguments{
\item{X}{The data matrix to perform clustering upon (items to cluster in rows).}

\item{K}{The number of components to model (upper limit on the number of clusters found).}

\item{labels}{Vector item labels to initialise from.}

\item{R}{The number of iterations to run for.}

\item{thin}{thinning factor for samples recorded.}

\item{concentration}{Vector of concentrations for mixture weights (recommended to be symmetric).}

\item{dataType}{Int, 0: independent Gaussians, 1: Multivariate normal, or 2: Categorical distributions.}
}
\value{
Named list of the matrix of MCMC samples generated (each row 
corresponds to a different sample) and BIC for each saved iteration.
}
\description{
Performs MCMC sampling for a mixture model with batch effects.
}
